+++ harmonica/filters/__init__.py
@@ -7,4 +7,12 @@
 """
 Frequency domain filters meant to be applied on regular grids
 """
-from ._filters import derivative_upward_kernel
+from ._filters import (
+    derivative_easting_kernel,
+    derivative_northing_kernel,
+    derivative_upward_kernel,
+    gaussian_highpass_kernel,
+    gaussian_lowpass_kernel,
+    reduction_to_pole_kernel,
+    upward_continuation_kernel,
+)